html/dom/elements/global-attributes/dataset-enumeration.html

<!DOCTYPE html>
<html>
<head>
<title>Dataset - Enumeration</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Dataset - Enumeration</h1>
<div id="log"></div>
<script>
function enumerate(dataset) {
  var result = [];
  for (var item in dataset) {
    result.push(item);
  }
  return result;
}

test(function() {
  var d = document.createElement("div");
  d.setAttribute('data-foo', "value");
  d.setAttribute('data-bar', "value");
  d.setAttribute('data-baz', "value");
  assert_array_equals(enumerate(d.dataset), ['foo', 'bar', 'baz']);
}, "A dataset should be enumeratable.");

test(function() {
  var d = document.createElement("div");
  d.setAttribute('data-foo', "value");
  d.setAttribute('data-bar', "value");
  d.setAttribute('dataFoo', "value");
  assert_array_equals(enumerate(d.dataset), ['foo', 'bar']);
}, "Only attributes who qualify as dataset properties should be enumeratable in the dataset.");
</script>
</body>
</html>
